make everything work with new doLayout.
This modifies all the contrib modules to work (so far as I know) with the
new contrib layout.  The exception is the LayoutHooks module, which isn't
used.  It exports an API that is inherently unsafe, so far as I can tell
(and always has been).

darcs-hash:20070623210952-72aca-1993ca13dc6996b59fedacc271c03fbaf87eabaa