RunInXTerm: a simple module to run commands in an X terminal
This is just a wrapper around spawn to launch commands in an X
terminal: runInXTerm will check the for the XTERMCMD environmental
variable, and use it to run the command. If the variable is not set
the command will be executed with xterm.

darcs-hash:20070807101603-32816-17799bba353c4b0fd5816842e8659512874d708f