add swapping capability in WindowNavigation.
This allows you to reorder your windows geometrically, by
swapping the currently focussed window with ones that are
up/down/right/left of it.  The idea is that we should be
able to manipulate windows based on the visual layout of
the screen rather than some (possibly obscure) logical ordering.

darcs-hash:20071003151755-72aca-7a83105557a561ae554110f1bc1617bed842ef12