(un)Manage Docks based on WINDOW_TYPE
Hi,

this is a replacement for the example code in Config.hs that should detect
and unamange, for example, the gnome-panel.

The problem with that code is that it also unamangs dialog boxes from gnome-panel
which then are not usable (no keyboard intput, at least here).

Greetings,
Joachim

darcs-hash:20071006132802-c9905-befdbe36f6f280d987dfb62702b7a6e05ce3306c