fix Actions.Wap doc
warp 1 1 has a comment claiming that this moves the cursor to the lower *left*, but if you look at the warpToWindow haddock, it says that 1 1 is actually lower *right* - as indeed it proved to do. This was annoying as it led me astray for a minute or so.

darcs-hash:20080629115504-f7719-817bf2cf8a6d63688cd170daf6775d5e187c4d7f