FlexibleResize - Resize from edge, don't move adjust at opposite edge
Ignore-this: 3c6c0748a4b0d14bd39bcb88f10aade6

When resizing other corners than bottom-right, instead of adjusting to even columns/rows on the opposite side to it the same way as if resizing was made from the bottom right.

Also add the possibility to add an area in the middle of an edge where only that edge is resized, not the closest corner.

darcs-hash:20090530185437-8978f-82de436312c45d105593603fd58f293056f91c11