ungrab-keyboard-before-action
Ignore-this: 5b8f3c286e8231a4d7ade2acbb2ae84a

If an action that requires the keyboard to be grabbed (e.g., launching dmenu),
it is a race when submapping the action as to whether the action will have
access to the keyboard or not. To fix this, the keyboard should be ungrabbed
before executing the action.

darcs-hash:20110515210312-24e3f-42b3c50a280ec90bec11cb48b57152d062c33b01