Escape dzen markup and remove xmobar tags from window titles by default.
Ignore-this: cf56bff752fbf78ea06d5c0cb755f615

The issue was that window titles, such as those set by, for example a browser,
could set the window title to display something like

   <action=malicious shell command>normal title</action>

Which could be executed by xmobar (or dzen).

This adds a ppTitleSanitize which does the above functions. This way when users
override ppTitle, the benefits are not lost.

Thanks to Raúl Benencia and Joachim Breitner for bringing this to my attention.

darcs-hash:20130708144813-1499c-0c3e284d3523c0694b9423714081761813bc1e89