X.L.Master: Add FixMaster layout modifier
Ignore-this: 82e9736853287f753248af41843ceb6b

This layout modifier is useful for the case if you desire to add a master
pane that has fixed width (it's fixed even if there is just one window
opened). Especially nice feature if you don't want to have too wide
terminal in a master pane.

The layout is implemented as an addition to Master layout, so it reuses
most of the code.

darcs-hash:20141220011339-1836e-543b0c692297779d51fd6f3f71e820366c977721