X.L.AvoidFloats more useful default settings
Ignore-this: af6535bace9219619858751eeeff6520
Changed default settings with the simple layout modifier. Instead of asking for a bool indicating if all windows should be avoided, no such bool is asked for. No windows are avoided by default. I think this will be a more useful default setting since it would be annoying if dialogue windows are avoided. The same functionality is possible with the advanced constructor. This will be easier for new users.

This will break configurations using the old module, but this will not be much of an issue since the module has not been added to the repo as of this writing.

darcs-hash:20150310212022-3948e-381395bd216296de1739fd5c2c22ed0bdfc85c9d